+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.0
 
 - Initial release
+- Kirby spawns randomly and tries to swallow you